+++ akamai/resource_akamai_dns_record.go
@@ -425,6 +425,19 @@ func dnsRecordTargetSuppress(k, old, new string, d *schema.ResourceData) bool {
 		compList = newTargetList
 	}
 
+	if recordtype == "AAAA" {
+		log.Printf("AAAA Suppress. baseval: [%v]", baseVal)
+		fullBaseval := FullIPv6(net.ParseIP(baseVal))
+		for _, compval := range compList {
+			log.Printf("AAAA Suppress. compval: [%v]", compval)
+			fullCompval := FullIPv6(net.ParseIP(compval.(string)))
+			if fullBaseval == fullCompval {
+				return true
+			}
+		}
+		return false
+	}
+
 	if recordtype == "CAA" {
 		basevalsplit := strings.Split(strings.Trim(baseVal, "\""), "\"")
 		baseVal = strings.Join(basevalsplit, "")
@@ -534,6 +547,29 @@ func getRecordLock(zone, host, recordtype string) *sync.Mutex {
 
 }
 
+func bumpSoaSerial(name string, d *schema.ResourceData, zone string, host string) (recordFunction, error) {
+	// Get SOA Record
+	recordset, e := dnsv2.GetRecord(zone, host, "SOA")
+	if e != nil {
+		return nil, fmt.Errorf("error looking up SOA record for %s: %s", host, e.Error())
+	}
+	rdataFieldMap := dnsv2.ParseRData("SOA", recordset.Target)
+	serial := rdataFieldMap["serial"].(int) + 1
+	d.Set("serial", serial)
+	newrecord, err := bindRecord(d)
+	if err != nil {
+		return nil, err
+	}
+	if name == "CREATE" {
+		return newrecord.Save, nil
+	} else if name == "UPDATE" {
+		return newrecord.Update, nil
+	}
+
+	return nil, fmt.Errorf("Bad Function")
+
+}
+
 // Record function signature
 type recordFunction func(string, ...bool) error
 
@@ -548,6 +584,16 @@ func executeRecordFunction(name string, d *schema.ResourceData, fn recordFunctio
 				log.Printf("[WARNING] [Akamai DNSv2] Concurrency Conflict")
 				opRetry -= 1
 				time.Sleep(100 * time.Millisecond)
+				e = fn(zone, rlock)
+				continue
+			} else if (name == "CREATE" || name == "UPDATE") && strings.Contains(e.(*dnsv2.RecordError).Error(), "SOA serial number must be incremented") {
+				log.Printf("[WARNING] [Akamai DNSv2] SOA Serial Number needs incrementing")
+				opRetry -= 1
+				time.Sleep(5 * time.Second) // let things quiesce
+				fn, err := bumpSoaSerial(name, d, zone, host)
+				if err != nil {
+					return err
+				}
 				e = fn(zone, rlock)
 				continue
 			} else if name == "DELETE" && e.(dnsv2.ConfigDNSError).NotFound() == true {
@@ -603,6 +649,15 @@ func resourceDNSRecordCreate(d *schema.ResourceData, meta interface{}) error {
 	// serialize record creates of same type
 	getRecordLock(zone, host, recordtype).Lock()
 	defer getRecordLock(zone, host, recordtype).Unlock()
+
+	/*
+		// works but serializes all recordset creates and updates per host
+		if recordtype != "SOA" {
+			// TF is multi threaded. Can't update SOA concurrently with other records
+			getRecordLock(zone, zone, "SOA").Lock()
+			defer getRecordLock(zone, zone, "SOA").Unlock()
+		}
+	*/
 
 	if recordtype == "SOA" {
 		// A default SOA is created automagically when the primary zone is created ...
@@ -854,6 +909,7 @@ func resourceDNSRecordRead(d *schema.ResourceData, meta interface{}) error {
 	}
 	log.Printf("[DEBUG] [Akamai DNSv2] READ record data read JSON %s", string(b1))
 	rdataFieldMap := dnsv2.ParseRData(recordtype, record.Target) // returns map[string]interface{}
+	targets := dnsv2.ProcessRdata(record.Target, recordtype)
 	if recordtype == "MX" {
 		// calc rdata sha from read record
 		sort.Strings(record.Target)
@@ -881,14 +937,28 @@ func resourceDNSRecordRead(d *schema.ResourceData, meta interface{}) error {
 				}
 			}
 		}
+	} else if recordtype == "AAAA" {
+		sort.Strings(record.Target)
+		rdataString := strings.Join(record.Target, " ")
+		shaRdata := getSHAString(rdataString)
+		if sha1hash == shaRdata {
+			// don't care if short or long notation
+			return nil
+		} else {
+			// could be either short or long notation
+			newrdata := make([]string, 0, len(record.Target))
+			for _, rcontent := range record.Target {
+				newrdata = append(newrdata, rcontent)
+			}
+			d.Set("target", newrdata)
+			targets = newrdata
+		}
 	} else {
 		// Parse Rdata. MX special
 		for fname, fvalue := range rdataFieldMap {
 			d.Set(fname, fvalue)
 		}
 	}
-
-	targets := dnsv2.ProcessRdata(record.Target, recordtype)
 	if len(targets) > 0 {
 		sort.Strings(targets)
 		if recordtype == "SOA" {
@@ -901,6 +971,9 @@ func resourceDNSRecordRead(d *schema.ResourceData, meta interface{}) error {
 					log.Printf("[DEBUG] [Akamai DNSv2] READ SOA RECORD CHANGE: SOA OK ")
 				}
 			}
+		} else if recordtype == "AKAMAITLC" {
+			extractTlcString := strings.Join(targets, " ")
+			sha1hash = getSHAString(extractTlcString)
 		}
 		d.Set("record_sha", sha1hash)
 		// Give terraform the ID
@@ -1971,7 +2044,7 @@ func checkSrvRecord(d *schema.ResourceData) error {
 		return err
 	}
 
-	if priority == 0 {
+	if priority < 0 || priority > 65535 {
 		return fmt.Errorf("Configuration argument priority must be set for SRV.")
 	}
 
@@ -2047,18 +2120,8 @@ func checkSoaRecord(d *schema.ResourceData) error {
 }
 
 func checkAkamaiTlcRecord(d *schema.ResourceData) error {
-	dnsname := d.Get("dns_name").(string)
-	answertype := d.Get("answer_type").(string)
 
-	if dnsname != "" {
-		return fmt.Errorf("Configuration argument dnsname is computed. It must not be set in AKAMAITLC.")
-	}
-
-	if answertype != "" {
-		return fmt.Errorf("Configuration argument answertype is computed. It must not be set in AKAMAITLC.")
-	}
-
-	return nil
+	return fmt.Errorf("AKAMAITLC is a READ ONLY record.")
 }
 
 func checkCaaRecord(d *schema.ResourceData) error {